Split out nested classes
